test(e2e): reuse shared startServer helper in NAA sample server
Extend the shared serverUtils.startServer with optional port/env params and a return value, and refactor the NAA sample server.js to reuse it instead of duplicating the spawn/logging logic. Keeps the two-app orchestration and coordinated shutdown local to the sample.


Copilot-Session: d9ccc5f1-36f3-4786-a416-e2afe65eb259
